Create new CSVExporter for accounting PnL report
