+++ rotkehlchen/tests/unit/test_history.py
@@ -6,66 +6,8 @@ from rotkehlchen.db.filtering import LedgerActionsFilterQuery
 from rotkehlchen.db.ledger_actions import DBLedgerActions
 from rotkehlchen.exchanges.data_structures import Trade
 from rotkehlchen.fval import FVal
-from rotkehlchen.history.events import limit_trade_list_to_period
 from rotkehlchen.history.types import HistoricalPriceOracle
 from rotkehlchen.types import Location, TradeType
-
-
-def test_limit_trade_list_to_period():
-    trade1 = Trade(
-        timestamp=1459427707,
-        location=Location.KRAKEN,
-        base_asset=A_ETH,
-        quote_asset=A_BTC,
-        trade_type=TradeType.BUY,
-        amount=FVal(1),
-        rate=FVal(1),
-        fee=FVal('0.1'),
-        fee_currency=A_ETH,
-        link='id1',
-    )
-    trade2 = Trade(
-        timestamp=1469427707,
-        location=Location.POLONIEX,
-        base_asset=A_ETH,
-        quote_asset=A_BTC,
-        trade_type=TradeType.BUY,
-        amount=FVal(1),
-        rate=FVal(1),
-        fee=FVal('0.1'),
-        fee_currency=A_ETH,
-        link='id2',
-    )
-    trade3 = Trade(
-        timestamp=1479427707,
-        location=Location.POLONIEX,
-        base_asset=A_ETH,
-        quote_asset=A_BTC,
-        trade_type=TradeType.BUY,
-        amount=FVal(1),
-        rate=FVal(1),
-        fee=FVal('0.1'),
-        fee_currency=A_ETH,
-        link='id3',
-    )
-
-    full_list = [trade1, trade2, trade3]
-    assert limit_trade_list_to_period(full_list, 1459427706, 1479427708) == full_list
-    assert limit_trade_list_to_period(full_list, 1459427707, 1479427708) == full_list
-    assert limit_trade_list_to_period(full_list, 1459427707, 1479427707) == full_list
-
-    expected = [trade2, trade3]
-    assert limit_trade_list_to_period(full_list, 1459427708, 1479427707) == expected
-    expected = [trade2]
-    assert limit_trade_list_to_period(full_list, 1459427708, 1479427706) == expected
-    assert limit_trade_list_to_period(full_list, 0, 10) == []
-    assert limit_trade_list_to_period(full_list, 1479427708, 1479427719) == []
-    assert limit_trade_list_to_period([trade1], 1459427707, 1459427707) == [trade1]
-    assert limit_trade_list_to_period([trade2], 1469427707, 1469427707) == [trade2]
-    assert limit_trade_list_to_period([trade3], 1479427707, 1479427707) == [trade3]
-    assert limit_trade_list_to_period(full_list, 1459427707, 1459427707) == [trade1]
-    assert limit_trade_list_to_period(full_list, 1469427707, 1469427707) == [trade2]
-    assert limit_trade_list_to_period(full_list, 1479427707, 1479427707) == [trade3]
 
 
 def test_query_ledger_actions(events_historian, function_scope_messages_aggregator):
